+++ Assets/Replicator/ParticleSystemCleanup.cs
@@ -4,6 +4,8 @@ namespace Replicator {
 	[AddComponentMenu("Pooling/Particle System Cleanup")]
 	[RequireComponent(typeof(ParticleSystem))]
 	public class ParticleSystemCleanup : MonoBehaviour, IPooled {
+		[SerializeField, Tooltip(Strings.ParticlesResetTooltip)]
+		private ParticleSystemStopBehavior resetBehaviour;
 		private new ParticleSystem particleSystem;
 
 		private void Start() {
@@ -11,7 +13,8 @@ namespace Replicator {
 		}
 
 		public void OnRecycle() {
-			particleSystem.Stop(true, ParticleSystemStopBehavior.StopEmittingAndClear);
+			particleSystem.Stop(true, resetBehaviour);
+			particleSystem.time = 0f;
 		}
 
 		public void OnSpawn() {